Replace Kaercher[TM] manual by Markdown version.
This should make it a lot easier to change and review stuff.

Along with the conversion from the ODT, I've fixed a few spelling errors
as well. Or to be more correct: Not adopted them in the first place.

The file name of the Markdown version is now ASCII-only as well as the
other documents.

Closes #165.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>